requirements.txt
flask == 3.1.2
gunicorn == 22.0.0
supervisor == 4.3.0
python-dotenv == 1.0.0
google-generativeai == 0.8.3